docgen/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
streamlit>=1.28.0
python-docx>=1.0.0
requests>=2.28.0

# Optional: richer in-app editor (bold, italic, lists). Without these, a plain text area is used.
# streamlit-quill>=0.2.0
# streamlit-lexical>=0.1.0
